+++ packages/cubejs-client-vue3/src/QueryBuilder.js
@@ -383,7 +383,7 @@ export default {
         this.chartType = chartType || this.chartType;
         this.pivotConfig = ResultSet.getNormalizedPivotConfig(
           validatedQuery,
-          pivotConfig || this.pivotConfig
+          pivotConfig !== undefined ? pivotConfig : this.pivotConfig
         );
         this.copyQueryFromProps(validatedQuery);
       }