+++ rust/cubestore/cubestore-sql-tests/src/tests.rs
@@ -1579,7 +1579,13 @@ async fn ilike(service: Box<dyn SqlClient>) {
         .exec_query(
             "INSERT INTO s.strings(t, pat) \
              VALUES ('aba', '%ABA'), ('ABa', '%aba%'), ('CABA', 'aba%'), ('ZABA', '%a%b%a%'), ('ZZZ', 'zzz'), ('TTT', 'TTT'),\
-             ('some_underscore', '%some\\\\_underscore%')",
+             ('some_underscore', '%some\\\\_underscore%'),\
+             ('test [ special 1', '%test [%'),\
+             ('test ( special 2', '%test (%'),\
+             ('111 test {)?*|+aaa', '%test {)?*|+aaa'),\
+             ('test2 }]\\\\222 ', 'test2 }]\\\\\\\\%')\
+             ",
+
         )
         .await
         .unwrap();
@@ -1615,6 +1621,17 @@ async fn ilike(service: Box<dyn SqlClient>) {
         .unwrap();
     assert_eq!(to_rows(&r), rows(&["some_underscore"]));
 
+    let r = service
+        .exec_query("SELECT t FROM s.strings WHERE t ILIKE CONCAT('%', '(', '%') ORDER BY t")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&["test ( special 2"]));
+
+    let r = service
+        .exec_query("SELECT t FROM s.strings WHERE t ILIKE CONCAT('%', '?*|+', '%') ORDER BY t")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&["111 test {)?*|+aaa"]));
     // Compare constant string with a bunch of patterns.
     // Inputs are: ('aba', '%ABA'), ('ABa', '%aba%'), ('CABA', 'aba%'), ('ZABA', '%a%b%a%'),
     //             ('ZZZ', 'zzz'), ('TTT', 'TTT').
@@ -1624,6 +1641,12 @@ async fn ilike(service: Box<dyn SqlClient>) {
         .unwrap();
     assert_eq!(to_rows(&r), rows(&["%ABA", "%a%b%a%", "%aba%", "aba%"]));
 
+    let r = service
+        .exec_query("SELECT pat FROM s.strings WHERE 'ggggtest (fjfj)' ILIKE pat ORDER BY pat")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&["%test (%"]));
+
     // Compare array against array.
     let r = service
         .exec_query("SELECT t, pat FROM s.strings WHERE t ILIKE pat ORDER BY t")
@@ -1632,12 +1655,16 @@ async fn ilike(service: Box<dyn SqlClient>) {
     assert_eq!(
         to_rows(&r),
         rows(&[
+            ("111 test {)?*|+aaa", "%test {)?*|+aaa"),
             ("ABa", "%aba%"),
             ("TTT", "TTT"),
             ("ZABA", "%a%b%a%"),
             ("ZZZ", "zzz"),
             ("aba", "%ABA"),
             ("some_underscore", "%some\\_underscore%"),
+            ("test ( special 2", "%test (%"),
+            ("test [ special 1", "%test [%"),
+            ("test2 }]\\222 ", "test2 }]\\\\%")
         ])
     );